backend/requirements.txt
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Markdown==3.5
psycopg[binary,c]==3.1.12
gunicorn==21.2.0
